+++ .nia/wiki/Changelog.md
@@ -2,4 +2,6 @@
 
 ## 2026-02-05
 
+- **[Testing]** Validation Test Issue (Issue #3)
+- **[Testing]** Validation Test Issue (Issue #2)
 - **[Testing]** Validation Test Issue (Issue #1)